make worker use "make list function"
this function isn't upstream yet but will be patched to all used image
builders (like package_list). this allows to determine the manifest in
advance to know the unique and reproducible manifest_hash before
creating the image. this results in signed images by the worker.

the previous "solution" was to rename the image and contents of the
sha256sums and then sign it again, *ugh*

Signed-off-by: Paul Spooren <mail@aparcar.org>