rework: support based on metadata
parse ImageBuilder output to see if and what devices are supported per
profile. This greatly simplifies the mapping between board_name and
profile.

Additionally check if "metadata" is added to the firmware images, this
is important to see if ucert/usign signatures are available. Goal is to
only support sysupgrades for devices supporting image verification.

Signed-off-by: Paul Spooren <mail@aparcar.org>