add /api/build/<request_hash> API
Now it is possible to directly request a known `request_hash`.

This should be used by all clients to run the verification function of
the server less often. Meaning a client fires the first request,
receives a `reuqest_hash` and can continue polling the hash.

Also usable for front ends to share static links of build images.

Signed-off-by: Paul Spooren <mail@aparcar.org>