fixup: empty package list and diff_packages name
Diffing an empty packages list does not make any sense and resulted in
an error with Redis pipeline. Now just ignore those empty packages
lists.

diff_packages is used when requesting images from a running machines, so
the server checks which packages are installed by default and removes
them actively if not already installed on the running device. Say `opkg`
is installed by default but not on the running system, `diff_packages`
adds `-opkg` to the packages list (once implemented).

Signed-off-by: Paul Spooren <mail@aparcar.org>