build: Fixup ImageBuiler download timezone
The timezones of remote and local files were not taken into account
resulting in frequent downloading of an up to date ImageBuilder.

Now just store whatever last-modified comes from the server and store it
in a `stamp` file. It's not important when the ImageBuiler changed but
only if. Whenever locally a different stamp is available, upgrade.

Signed-off-by: Paul Spooren <mail@aparcar.org>