config: add MAPPING_ABI option
This options allows to manually set an abi mapping. Problem is that
packages names on 19.07 releases contain an ABI version. The ABI version
however may change between releases so that the server won't find the
package anymore. An example would be `libubus2017` is called
`lububus2018` in the next release, resulting in `libubus2017` to be an
unknown package in the latter release.

To solve this `MAPPING_ABI` can be defined to map e.g. `libubus2017` to
the *ABI less* package name `libubus`. The package manager `opkg` will
automatically select the latest `libubus` version.

In releases post 19.07 this is automatically done on device since the
attached ABI version is separately stored on device.

Signed-off-by: Paul Spooren <mail@aparcar.org>