api: prefer profile directly over mapped one
Within OpenWrt the profile does not always matches the `board_name`,
which is determined via device tree files. As a workaround ASU
automatically parsed the "supported" devices and mapped incomming
requests to profiles. In some rare cases (#319) the support profile is
actually a different profile than the one installed, in this case a
device that support images with and without nand storage.

To fix this, first check if a profile with the device tree model exists
(by replacing `,` with `_`) and only use the *supported mapping* if it
does not.

Fixes: #319

Signed-off-by: Paul Spooren <mail@aparcar.org>