asu: allow the usage of branches.yml
It's easier to write and understand. It's now possible to define a file
called `branches.yml` containing whatever branch information you'd like.

It's loaded from the config option BRANCHES_PATH if defined, if not the
internal branches.yml file is used.

To stay compatible with existing setups changes only happen if BRANCHES
wasn't previously defined, like it would be the case if still added in
the config.py

Signed-off-by: Paul Spooren <mail@aparcar.org>