api: drop server side handling of package_changes
This causes more issues than good. The server must only provide such
information and the client should request the user what do to or may
"auto default" to what `package_changes` suggest.

Changing this on the server side confuses clients that compare the
requested list of packages with the resulting list of packages.

Signed-off-by: Paul Spooren <mail@aparcar.org>