api: better handling of defaults
If defaults are not allowed but the parameter is supplied, don't error
out but instead just ignore it.

Always include the "defaults" content to the request hash, even if
empty to avoid an extra case, adjust the corresponding hashes.

If defaults are set, remove the created package after an hour.

Signed-off-by: Paul Spooren <mail@aparcar.org>