config: split 'upstream_url' to internal and external values
Attempt to mitigate lag in CDN update which causes 'downloads'
site to lag behind actual builds, and appears to propagate into the
ASU server's metatdata.  We do this by splitting the 'upstream_url'
into a "local" and "CDN" value, the former is used for all internal
uses and the latter is reported to clients.

See https://github.com/openwrt/asu/issues/853#issuecomment-2265311030
for more detail.

Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>