util: improve package file parsing
Occasionally, upstream package files are being terminated with extra newlines,
resulting in '"null": null' entries in the resulting package index.  Strip the
input text to avoid this.

Change the splitting method on the text to both reduce memory footprint and
speed things up.

Add 'headersonly' to parser to speed it up by about 10%.

Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>