update: include any upcoming version in latest release list
The LuCI Attended Sysupgrade app uses the 'latest' list in overview.json to
determine versions shown to the end user.  It currently does not show the
'upcoming_version', which means users don't see the 24.10.0-rc series.

If 'upcoming_version' exists, then let's add it to that list, so users can
test out the RCs.

Fixes: https://github.com/openwrt/openwrt/issues/17595
Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>