api: when specific build request failures occur attempt to recover
If a build request has been received for a version, branch, target or subtarget
that somehow has missed an update via webhooks, we invoke the 'update' mechanism
and proceed with the error processing.

The result is that this first build request still fails, but any future requests
on the branch/target will (hopefully) succeed.

Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>