utils: set cache timeout on upstream data
The versions, targets and subtargets were being permanently cached, so when
say upstream .versions.json changed, that change was not being picked up and
propagated into the overview.json.

Added ttl to the cache entries, defaulting to one hour, so that the various
data will not be out-of-date for long.

Authored-by: @dc366
Links: https://github.com/openwrt/asu/issues/1327#issuecomment-2808679592
Fixes: #1327 #1330
Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>