build: put a limit on the number of queued build requests
Restrict the number of build requests that may be in the queue by returning
status=529 (server overloaded).

From observing the rate at which jobs are processed on the production server,
it appears they are cleared at an average of just over 5 jobs/minute, so the
selected queue length limit of 200 would equate to about a 40 minute wait.

Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>
Signed-off-by: Paul Spooren <mail@aparcar.org>