util: handle None returned as output string
In newer python podman versions, the exec_run method may return
None as a value for either of stdout or stderr when there is no
output, where they previously always returned an empty bytestring.
This raises an AttributeError when it occurs cascading into an
"Internal server error", fix it by checking before decoding the
bytestrings.

Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>