util: handle missing .versions.json gracefully
On rare occasion the download of .versions.json fails with a 404,
probably due to requesting it while it's being updated.  If this
happens during a build request, it causes an "Internal server error"
and the job request fails.  Fix that by simply not updating the
internal versions list and continuing on until the versions are
available.

Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>