build: handle errors during container download
If a container image is missing for whatever reason, then the image
pull raises an exception:

 ImageNotFound: 404 Client Error: Not Found ({"message":"manifest unknown"}

This is propagated to the ASU clients as a "status 500 - Internal
server error".  The ASU clients then simply report "init" as an
error with no supporting details, leaving the user completely in
the dark as to what is going on.

We add an error handler around the image pull, so now the server
reports very specific details about the failure, helping diagnose
the issue:

 RuntimeError: Image not found: ghcr.io/openwrt/imagebuilder:lantiq-xrx200-v24.10.1

Link: https://forum.openwrt.org/t/luci-attended-sysupgrade-support-thread/230552/68
Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>