util: handle missing .targets.json file
During the creation of a new maintenance release, it is possible
that the new version's .target.json file has not yet been created.
This causes a "500 internal server" error for all users, even if they
request a different version.  Rather than crash attempting to decode
the 404 html code as json, we simply return an empty dictionary.

This results in users seeing the "Unsupported target" message if
they try to build the new version, but that's better than crashing
for everyone.

Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>