build-request: sanitize profile before use
Older versions of the LuCI ASU app did not sanitize the device
profile by replacing commas with underscores, it simply inserted
the raw output from 'ubus call system board' into the request.
This causes the build request to be rejected by the ASU server:

    Unsupported profile: sinovoip,bananapi-m2-berry. The requested
        profile was either dropped or never existed...

Let's make sure to sanitize profiles as soon as we receive them,
to avoid this issue and ensure that all of our use is consistent
with the upstream profile lists.

Link: https://forum.openwrt.org/t/luci-attended-sysupgrade-support-thread/230552/246
Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>