feat(error-log): add a concise error logging capability
Record build errors in a rotating log to allow easier diagnosis
of upstream imagebuilder and package build issues.  Add a separate
web page for access to the log's content.  Log content is exposed
via an API call, so other downstream tools may utilize the data.

Note that the log content is intentionally minimal and anonymized
to minimize any use for tracking ASU users.

Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>