stats: add a builds-by-branch chart to show trends
Show a time series plot of builds by branch.  After playing with
various values, I decided upon a 7-day bucket and a 6-month total
interval, which gives a long enough period to get a good feel for
trends without being too dense and confusing.

I also aggregated the data for all versions within a branch by
default (so, 24.10.* all appear as just "24.10"), but the API
call has a "branch" parameter, which allows you to show just the
versions within a given branch.  I did not expose this in the UI,
let's make sure this works first.

Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>