util: eliminate extra 'upcoming_version' in versions list
Do major code cleanup on 'reload_versions' with two fixes:

1) We were adding the 'upcoming_version' to the versions list twice
if it was already in the json's 'versions' list.  This was mostly
aesthetic and didn't break anything, but it looks bad in ASU clients
to have it twice, so let's clean it up.

2) The code wasn't checking if the branch was enabled, and just
adding a version even if it was in a no longer supported branch.

Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>