util: properly detect version at which kmod split occurred
A sloppy test for the version at which the kmods directory was
split from the platform packages was string based and stopped at
version 24.  This caused requests for version 25 platform packages
to be incomplete, as there were no kmods included in the response.

Fix this by doing proper comparisons for major version values.

Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>